feat: alternative typing strategy
This adds a `src/near` directory with all near-related things and a
`src/near/contracts` sub-directory which exports contract instances from
its index file. This ends up having similar DevX throughout the rest of
the app, but keeps the contents of each file in `src/near/contracts`
more auto-generateable.
